+++ assets/javascripts/kitten/components/layout/dashboard-layout/stories.js
@@ -38,6 +38,11 @@ import {
   useWindowWidth,
   useDeepCompareEffect,
   TextCopy,
+  FlexWrapper,
+  RewardSummaryCard,
+  useRewardSummaryCardResizeObserver,
+  DropdownMenu,
+  EllipsisIcon,
 } from '../../..'
 
 import { Default as Table } from '../../organisms/tables/list-table/list-table.stories.js'
@@ -159,7 +164,7 @@ export default {
     selectedView: {
       name: 'Selected View (story prop)',
       control: 'radio',
-      options: ['flow', 'dashboard', 'table'],
+      options: ['flow', 'dashboard', 'table', 'rewards'],
     },
     displayHeader: { name: 'displayHeader (story prop)', control: 'boolean' },
     displayAlerts: { name: 'displayAlerts (story prop)', control: 'boolean' },
@@ -248,10 +253,10 @@ export const Default = ({
                 Contributeurs
               </DashboardMenu.Item>
               <DashboardMenu.Item
-                href="?id=layout-dashboardlayout--default&args=selectedView:table&viewMode=story&custom=a"
+                href="?id=layout-dashboardlayout--default&args=selectedView:rewards&viewMode=story&custom=a"
                 isActive={
                   window.location.search ===
-                  '?id=layout-dashboardlayout--default&args=selectedView:table&viewMode=story&custom=a'
+                  '?id=layout-dashboardlayout--default&args=selectedView:rewards&viewMode=story&custom=a'
                 }
                 icon={() => <TagIcon color="currentColor" />}
                 onClick={() => closeSideBar()}
@@ -373,6 +378,7 @@ export const Default = ({
       {selectedView === 'flow' && <FlowExample {...flowProps} />}
       {selectedView === 'dashboard' && <DashExample />}
       {selectedView === 'table' && <TableExample />}
+      {selectedView === 'rewards' && <RewardsExample />}
     </DashboardLayout>
   )
 }
@@ -412,6 +418,156 @@ const TableExample = () => (
     </div>
   </>
 )
+
+const RewardsExample = () => {
+  const { ref, size } = useRewardSummaryCardResizeObserver()
+
+  return (
+    <div ref={ref}>
+      <Title
+        modifier="quinary"
+        className="k-u-margin-none k-u-margin-bottom-triple"
+      >
+        Curabitur blandit tempus porttitor.
+      </Title>
+
+      <FlexWrapper gap={15}>
+        <RewardSummaryCard.TitleBar
+          values={{
+            image: 'Visuel de la contrepartie',
+            title: 'Titre de la contrepartie',
+            amount: 'Montant',
+            contributions: 'Contributions',
+            availability: 'Disponibilités',
+          }}
+          className="k-u-hidden@xs-down k-u-margin-bottom-double"
+          id="RewardSummaryList"
+          size={size}
+        />
+
+        <RewardSummaryCard size={size}>
+          <RewardSummaryCard.Image>
+            <img alt="" src="/kitten.jpg" />
+          </RewardSummaryCard.Image>
+          <div style={{ display: 'grid' }}>
+            <RewardSummaryCard.Title aria-describedby="RewardSummaryList-title">
+              Stickers Free Boobs Club  Stickers Free Boobs Club  Stickers Free Boobs Club.
+            </RewardSummaryCard.Title>
+            <RewardSummaryCard.TitleTag text="Contrepartie star" />
+          </div>
+          <RewardSummaryCard.Amount aria-describedby="RewardSummaryList-amount">
+            10 000€
+          </RewardSummaryCard.Amount>
+          <RewardSummaryCard.Contribution aria-describedby="RewardSummaryList-contribution">
+            1 468 000
+          </RewardSummaryCard.Contribution>
+          <RewardSummaryCard.Availability aria-describedby="RewardSummaryList-availability">
+            Illimitée
+          </RewardSummaryCard.Availability>
+          <RewardSummaryCard.Options>
+            <DropdownMenu
+              button={() => <EllipsisIcon color={COLORS.font1} />}
+              menuPosition="left"
+            >
+              <DropdownMenu.Link href="#">Editer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Dupliquer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Désactiver</DropdownMenu.Link>
+              <DropdownMenu.Link>Supprimer</DropdownMenu.Link>
+            </DropdownMenu>
+          </RewardSummaryCard.Options>
+        </RewardSummaryCard>
+
+        <RewardSummaryCard size={size}>
+          <div style={{ display: 'grid' }}>
+            <RewardSummaryCard.Title aria-describedby="RewardSummaryList-title">
+              Stickers Free Boobs Club
+            </RewardSummaryCard.Title>
+            <RewardSummaryCard.TitleTag text="Contrepartie" />
+          </div>
+          <RewardSummaryCard.Amount aria-describedby="RewardSummaryList-amount">
+            10 000€
+          </RewardSummaryCard.Amount>
+          <RewardSummaryCard.Contribution aria-describedby="RewardSummaryList-contribution">
+            1 468 000
+          </RewardSummaryCard.Contribution>
+          <RewardSummaryCard.Availability aria-describedby="RewardSummaryList-availability">
+            Illimitée
+          </RewardSummaryCard.Availability>
+          <RewardSummaryCard.Options>
+            <DropdownMenu
+              button={() => <EllipsisIcon color={COLORS.font1} />}
+              menuPosition="left"
+            >
+              <DropdownMenu.Link href="#">Editer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Dupliquer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Désactiver</DropdownMenu.Link>
+              <DropdownMenu.Link>Supprimer</DropdownMenu.Link>
+            </DropdownMenu>
+          </RewardSummaryCard.Options>
+        </RewardSummaryCard>
+
+        <RewardSummaryCard size={size}>
+          <RewardSummaryCard.Image>
+            <img alt="" src="/kitten.jpg" />
+          </RewardSummaryCard.Image>
+          <div style={{ display: 'grid' }}>
+            <RewardSummaryCard.Title aria-describedby="RewardSummaryList-title">
+              Stickers Free Boobs Club
+            </RewardSummaryCard.Title>
+            <RewardSummaryCard.TitleTag text="Contrepartie star" />
+          </div>
+          <RewardSummaryCard.Amount aria-describedby="RewardSummaryList-amount">
+            10 000€
+          </RewardSummaryCard.Amount>
+          <RewardSummaryCard.Contribution aria-describedby="RewardSummaryList-contribution">
+            1 468 000
+          </RewardSummaryCard.Contribution>
+          <RewardSummaryCard.Availability aria-describedby="RewardSummaryList-availability">
+            Illimitée
+          </RewardSummaryCard.Availability>
+          <RewardSummaryCard.Options>
+            <DropdownMenu
+              button={() => <EllipsisIcon color={COLORS.font1} />}
+              menuPosition="left"
+            >
+              <DropdownMenu.Link href="#">Editer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Dupliquer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Désactiver</DropdownMenu.Link>
+              <DropdownMenu.Link>Supprimer</DropdownMenu.Link>
+            </DropdownMenu>
+          </RewardSummaryCard.Options>
+        </RewardSummaryCard>
+        <RewardSummaryCard size={size}>
+          <div style={{ display: 'grid' }}>
+            <RewardSummaryCard.Title aria-describedby="RewardSummaryList-title">
+              Stickers Free Boobs Club
+            </RewardSummaryCard.Title>
+          </div>
+          <RewardSummaryCard.Amount aria-describedby="RewardSummaryList-amount">
+            10 000€
+          </RewardSummaryCard.Amount>
+          <RewardSummaryCard.Contribution aria-describedby="RewardSummaryList-contribution">
+            1 468 000
+          </RewardSummaryCard.Contribution>
+          <RewardSummaryCard.Availability aria-describedby="RewardSummaryList-availability">
+            Illimitée
+          </RewardSummaryCard.Availability>
+          <RewardSummaryCard.Options>
+            <DropdownMenu
+              button={() => <EllipsisIcon color={COLORS.font1} />}
+              menuPosition="left"
+            >
+              <DropdownMenu.Link href="#">Editer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Dupliquer</DropdownMenu.Link>
+              <DropdownMenu.Link href="#">Désactiver</DropdownMenu.Link>
+              <DropdownMenu.Link>Supprimer</DropdownMenu.Link>
+            </DropdownMenu>
+          </RewardSummaryCard.Options>
+        </RewardSummaryCard>
+      </FlexWrapper>
+    </div>
+  )
+}
 
 const FlowExample = ({
   flowLoading,
@@ -541,7 +697,7 @@ const FlowExample = ({
       ) : flowShowUnsavedText ? (
         <Text
           weight="light"
-          color="font2"
+          cssColor={COLORS.font3}
           size="tiny"
           className="k-u-hidden@xs-down"
           style={{ alignSelf: 'center' }}