+++ assets/javascripts/kitten/components/molecules/cards/reward-summary-card/components/options.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import classNames from 'classnames'
+
+export const Options = ({ className, ...props }) => {
+  return (
+    <div
+      className={classNames(
+        'k-RewardSummaryCard__options',
+        'k-u-hidden@xs-down',
+        className,
+      )}
+      {...props}
+    />
+  )
+}