+++ src/esm/components/molecules/cards/simple-card/components/image.js
@@ -17,7 +17,7 @@ export var Image = function Image(_ref) {
       style = _ref.style;
   return /*#__PURE__*/React.createElement("div", {
     className: classNames('k-SimpleCard__imageContainer', className, {
-      'k-SimpleCard__imageContainer--ratio': imageContainerRatio
+      'k-SimpleCard__imageContainer--ratio': !!imageContainerRatio
     }),
     style: _extends({}, style, {
       '--SimpleCard-image-container-background': imageContainerBackground,